+++ cvat/apps/engine/annotation.py
@@ -1090,12 +1090,19 @@ class TrackManager(ObjectManager):
             step = np.subtract(shape1["points"], shape0["points"]) / distance
             for frame in range(shape0["frame"] + 1, shape1["frame"]):
                 off = frame - shape0["frame"]
-                points = shape0["points"] + step * off
+                if shape1["outside"]:
+                    points = np.asarray(shape0["points"]).reshape(-1, 2)
+                else:
+                    points = (shape0["points"] + step * off).reshape(-1, 2)
                 shape = copy.deepcopy(shape0)
-                broken_line = geometry.LineString(points.reshape(-1, 2)).simplify(0.05, False)
+                if len(points) == 1:
+                    shape["points"] = points.flatten()
+                else:
+                    broken_line = geometry.LineString(points).simplify(0.05, False)
+                    shape["points"] = [x for p in broken_line.coords for x in p]
+
                 shape["keyframe"] = False
                 shape["frame"] = frame
-                shape["points"] = [x for p in broken_line.coords for x in p]
                 shapes.append(shape)
             return shapes
 
@@ -1248,6 +1255,9 @@ class TaskAnnotation:
                 ("flipped", str(db_task.flipped)),
                 ("created", str(timezone.localtime(db_task.created_date))),
                 ("updated", str(timezone.localtime(db_task.updated_date))),
+                ("start_frame", str(db_task.start_frame)),
+                ("stop_frame", str(db_task.stop_frame)),
+                ("frame_filter", db_task.frame_filter),
 
                 ("labels", [
                     ("label", OrderedDict([
@@ -1286,6 +1296,8 @@ class TaskAnnotation:
                 ("width", str(im_meta_data[0]["width"])),
                 ("height", str(im_meta_data[0]["height"]))
             ])
+            # Add source to dumped file
+            meta["source"] = str(db_task.video.path)
 
         with open(file_path, "w") as dump_file:
             dumper = XmlAnnotationWriter(dump_file)
@@ -1407,7 +1419,7 @@ class TaskAnnotation:
                             self._flip_shape(shape, im_w, im_h)
 
                         dump_data = OrderedDict([
-                            ("frame", str(shape["frame"])),
+                            ("frame", str(db_task.start_frame + shape["frame"] * db_task.get_frame_step())),
                             ("outside", str(int(shape["outside"]))),
                             ("occluded", str(int(shape["occluded"]))),
                             ("keyframe", str(int(shape["keyframe"])))