+++ cvat/apps/engine/static/engine/js/shapeCreator.js
@@ -54,12 +54,12 @@ class ShapeCreatorModel extends Listener {
             });
         }
 
-        if (this._defaultMode === 'interpolation' && this._defaultType === 'box') {
+        // FIXME: In the future we have to make some generic solution
+        if (this._defaultMode === 'interpolation' && ['box', 'points'].includes(this._defaultType)) {
             data.shapes = [];
             data.shapes.push(Object.assign({}, result, data));
-            this._shapeCollection.add(data, `interpolation_box`);
-        }
-        else {
+            this._shapeCollection.add(data, `interpolation_${this._defaultType}`);
+        } else {
             Object.assign(data, result);
             this._shapeCollection.add(data, `annotation_${this._defaultType}`);
         }
@@ -213,11 +213,14 @@ class ShapeCreatorView {
         }
 
         this._typeSelector.on('change', (e) => {
-            let type = $(e.target).prop('value');
-            if (type != 'box' && this._modeSelector.prop('value') != 'annotation') {
+            // FIXME: In the future we have to make some generic solution
+            const mode = this._modeSelector.prop('value');
+            const type = $(e.target).prop('value');
+            if (type !== 'box' && !(type === 'points' && this._polyShapeSize === 1)
+                && mode !== 'annotation') {
                 this._modeSelector.prop('value', 'annotation');
                 this._controller.setDefaultShapeMode('annotation');
-                showMessage('Poly shapes available only like annotation shapes');
+                showMessage('Only the annotation mode allowed for the shape');
             }
             this._controller.setDefaultShapeType(type);
         }).trigger('change');
@@ -227,20 +230,30 @@ class ShapeCreatorView {
         }).trigger('change');
 
         this._modeSelector.on('change', (e) => {
-            let mode = $(e.target).prop('value');
-            if (mode != 'annotation' && this._typeSelector.prop('value') != 'box') {
+            // FIXME: In the future we have to make some generic solution
+            const mode = $(e.target).prop('value');
+            const type = this._typeSelector.prop('value');
+            if (mode !== 'annotation' && !(type === 'points' && this._polyShapeSize === 1)
+                && type !== 'box') {
                 this._typeSelector.prop('value', 'box');
                 this._controller.setDefaultShapeType('box');
-                showMessage('Only boxes available like interpolation shapes');
+                showMessage('Only boxes and single point allowed in the interpolation mode');
             }
             this._controller.setDefaultShapeMode(mode);
         }).trigger('change');
 
         this._polyShapeSizeInput.on('change', (e) => {
             e.stopPropagation();
-            let size = + e.target.value;
+            let size = +e.target.value;
             if (size < 0) size = 0;
             if (size > 100) size = 0;
+            const mode = this._modeSelector.prop('value');
+            const type = this._typeSelector.prop('value');
+            if (mode === 'interpolation' && type === 'points' && size !== 1) {
+                showMessage('Only single point allowed in the interpolation mode');
+                size = 1;
+            }
+
             e.target.value = size || '';
             this._polyShapeSize = size;
         }).trigger('change');
@@ -265,6 +278,7 @@ class ShapeCreatorView {
             let size = this._polyShapeSize;
             let sizeDecrement = function() {
                 if (!--size) {
+                    numberOfPoints = this._polyShapeSize;
                     this._drawInstance.draw('done');
                 }
             }.bind(this);
@@ -323,7 +337,7 @@ class ShapeCreatorView {
                         this._drawInstance.draw('point', e);
                         lastPoint = {
                             x: e.clientX,
-                            y: e.clientY
+                            y: e.clientY,
                         };
                     }
                 }